+++ tests/integration/test_cloud_download.py
@@ -1,5 +1,8 @@
 import logging
+import operator as op
+import os
 import shutil
+from collections.abc import Callable
 from pathlib import Path
 
 import earthaccess
@@ -117,8 +120,16 @@ def test_earthaccess_can_download_cloud_collection_granules(tmp_path, daac):
             )
 
 
-@pytest.mark.parametrize("force", [True, False])
-def test_multi_file_granule(tmp_path, force):
+@pytest.mark.parametrize(
+    ("force", "cmp"),
+    [
+        # Redownloading should update all of the mtimes, so first list of mtimes will be less than second list of mtimes
+        (True, op.lt),
+        # No forced downloading, so first list of mtimes will be the same as the second list of mtimes
+        (False, op.eq),
+    ],
+)
+def test_multi_file_granule(tmp_path, force: bool, cmp: Callable[[float, float], bool]):
     # Ensure granules that contain multiple files are handled correctly
     granules = earthaccess.search_data(short_name="HLSL30", count=1)
     assert len(granules) == 1
@@ -127,17 +138,11 @@ def test_multi_file_granule(tmp_path, force):
     files = earthaccess.download(granules, str(tmp_path))
     assert {Path(f).name for f in urls} == {Path(f).name for f in files}
 
-    # Verify force behavior
+    # Make sure no temp files are left behind
+    assert len(os.listdir(tmp_path)) == len(files)
+
+    # Verify force behavior by calling download again and checking mtimes
     first_mtimes = [f.stat().st_mtime for f in files]
     second_files = earthaccess.download(granules, str(tmp_path), force=force)
     second_mtimes = [f.stat().st_mtime for f in second_files]
-    if force:
-        # Redownloading should update all of the mtimes
-        assert all(
-            mtime1 < mtime2 for mtime1, mtime2 in zip(first_mtimes, second_mtimes)
-        )
-    else:
-        # No forced downloading, so no change in any mtimes.
-        assert all(
-            mtime1 == mtime2 for mtime1, mtime2 in zip(first_mtimes, second_mtimes)
-        )
+    assert all(cmp(*mtime_pair) for mtime_pair in zip(first_mtimes, second_mtimes))